fix(StatusSeedPhraseInput): ensure `doneInsertingWord` is emitted
When `completeWorldFill()` was called, we relied on it to implicitly
emit `doneInsertingWord()` via the seed phrase input's text change
event.

However, it turns out `doneInsertingWord` is only emitted by the
component when the filtered suggestion list has one item.

In practice this isn't always the case, for example the word "fun"
might be desired, in which case the suggestion list includes "fun"
and "funny".

This introduces some third order effects, such that in Status Desktop
the view doesn't get notified that a word has been inserted, which then
causes it invalidate the mnemonic (or rather, it keeps the confirm
button disabled because it thinks there were less words inserted than
required).

To account for this we always explicitly emit the `doneInsertWord`
signal.